Add files via upload
This is the first thing I ever made all by myself.
We generate a random number between 0 and 1,000,000 and run if tests to see if it's even or divisible by 3, 4 or 5. It should blink when the number is prime, but I haven't seen it work.
I used the true random number generator on the ESP32 and ESP Servo library to get the buzzers to work. Those numbers get crazy without the 1M limit.